requirements.txt
numpy
gradio
pillow
tensorflow